feat(main_page): use new cubit
